refactor wave into isolated component
